requirements.txt
numpy
pandas
opencv-python
scikit-learn
torch
torchvision
matplotlib
seaborn
requests
Pillow
